requirements.txt
langchain_google_genai
streamlit
python-dotenv
google-generativeai
PyPDF2
langchain_community